Fix 2 or 3 bugs about cookie handling
* expires date is not an integer but a date formatted given some RFC
like 822...
* There is no need to re-set a cookie at every connexion: once is
enough
* It's worse when you set a different cookie every time because:
one is not able to retrieve a path and expires time from a cookie sent
by the client. so every time a new cookie is set, it's without an expires
and path so it's default to end of session and current path. So you end
up with a lot of cookies !

So this commit change the code to:
- only set once the cookie
- got correct expires date